+++ CHANGES.md
@@ -4,6 +4,8 @@ Version 1.2.0, 2020-03-TODO
 
 * Changed installation to fail when attempting to install on Python earlier
   than 3.5.
+* Improved API:
+  * `SourceAnalysis.state` is now a proper enum instead of a string.
 * Cleaned up project:
   * Changed continuous integration from Travis CI to Github actions in the hope
     that the CI build does not automatically break after a while because